\name{LSPM-package}
\Rdversion{1.1}
\alias{LSPM-package}
\alias{LSPM}
\docType{package}
\title{
  Leverage Space Portfolio Modeler
}
\description{
  Functions to implement Ralph Vince's Leverage Space Portfolio Model in R.
}
\details{
  \tabular{ll}{
    Package: \tab LSPM\cr
    Type: \tab Package\cr
    Version: \tab 0.1-0\cr
    Date: \tab 2009-09-20\cr
    License: \tab GPL Version 3 or later\cr
    LazyLoad: \tab yes\cr
  }
}
\author{
  Soren MacBeth, Joshua Ulrich, and Ralph Vince

  Maintainer: Joshua Ulrich <josh.m.ulrich@gmail.com>
}
\references{
  Ralph Vince, The Handbook of Portfolio Mathematics\cr
  Ralph Vince, The Leverage Space Trading Model
}
\keyword{ package }
%\seealso{
%~~ Optional links to other man pages, e.g. ~~
%~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
%}
%\examples{
%~~ simple examples of the most important functions ~~
%}
