\name{port}
\alias{port}
\docType{data}
\title{ Sample lsp portfolio data }
\description{
  Multiple strategy example data found on pp. 84-87, 169 of
  \emph{The Leverage Space Trading Model} by Ralph Vince.
}
\usage{
  data(port)
}
\format{ A \code{lsp} class object containing 12 observations. }
\source{ The Leverage Space Trading Model, pp 84-87, 169 }
\references{
  Vince, Ralph (2009) \emph{The Leverage Space Trading Model}.
  New York: John Wiley & Sons, Inc.
}
%\examples{
%  data(ttrc)
%  plot(tail(ttrc[,"Close"],100), type="l")
%}
\keyword{ datasets }
